soa/requirements.txt
faust-streaming
